pyproject.toml
[project]
name = "distributed-trip-level-energy-consumption-forecasting"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.12",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
